extract machine-independent code
