+++ include/kernel/infrastructure/acpi/asm/acpi.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024-2025 Philippe Aubertin.
+ * Copyright (C) 202-2025 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
@@ -29,32 +29,18 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_ACPI_TYPES_H
-#define JINUE_KERNEL_INFRASTRUCTURE_ACPI_TYPES_H
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_ACPI_ASM_ACPI_H
+#define JINUE_KERNEL_INFRASTRUCTURE_ACPI_ASM_ACPI_H
 
-#include <stdint.h>
+#define ACPI_V1_REVISION    0
 
-typedef struct {
-    uint64_t addr;
-    uint64_t size;
-    uint32_t type;
-} acpi_addr_range_t;
+#define ACPI_V2_REVISION    2
 
-typedef struct {
-    char        signature[4];
-    uint32_t    length;
-    uint8_t     revision;
-    uint8_t     checksum;
-    char        oemid[6];
-    char        oem_table_id[8];
-    uint32_t    oem_revision;
-    uint32_t    creator_id;
-    uint32_t    creator_revision;
-} acpi_table_header_t;
+#define ACPI_V1_RSDP_SIZE   20
 
-typedef struct {
-    const char   *signature;
-    const void  **ptr;
-} acpi_table_def_t;
+/* Arbitrary value expected to be large enough to accomodate any real table
+ * while ensuring we don't create arbitrary large mappings because of garbage
+ * data in length members. */
+#define ACPI_TABLE_MAX_SIZE (128 * 1024)
 
 #endif
