+++ include/kernel/infrastructure/acpi/tables.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024-2025 Philippe Aubertin.
+ * Copyright (C) 2025 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
@@ -29,32 +29,42 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_ACPI_TYPES_H
-#define JINUE_KERNEL_INFRASTRUCTURE_ACPI_TYPES_H
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_ACPI_TABLES_H
+#define JINUE_KERNEL_INFRASTRUCTURE_ACPI_TABLES_H
 
+#include <kernel/infrastructure/acpi/types.h>
 #include <stdint.h>
 
 typedef struct {
-    uint64_t addr;
-    uint64_t size;
-    uint32_t type;
-} acpi_addr_range_t;
-
-typedef struct {
-    char        signature[4];
-    uint32_t    length;
-    uint8_t     revision;
+    char        signature[8];
     uint8_t     checksum;
     char        oemid[6];
-    char        oem_table_id[8];
-    uint32_t    oem_revision;
-    uint32_t    creator_id;
-    uint32_t    creator_revision;
-} acpi_table_header_t;
+    uint8_t     revision;
+    uint32_t    rsdt_address;
+    uint32_t    length;
+    uint64_t    xsdt_address;
+    uint8_t     extended_checksum;
+    uint8_t     reserved[3];
+} acpi_rsdp_t;
 
 typedef struct {
-    const char   *signature;
-    const void  **ptr;
-} acpi_table_def_t;
+    acpi_table_header_t header;
+    uint32_t            entries[];
+} acpi_rsdt_t;
+
+typedef struct {
+    acpi_table_header_t header;
+    /* TODO */
+} acpi_fadt_t;
+
+typedef struct {
+    acpi_table_header_t header;
+    /* TODO */
+} acpi_madt_t;
+
+typedef struct {
+    acpi_table_header_t header;
+    /* TODO */
+} acpi_hpet_t;
 
 #endif
